refactor: abstract search common ui out
